requirements.txt
PySide6
pycryptodome
smartcard
